+++ src/leds.h
@@ -1,7 +1,8 @@
 #include <stdint.h>
+
 void Leds_Create(uint16_t *direccion);
 
-void Leds_On(uint8_t led);
-void Leds_Off(uint8_t led);
+void Leds_On(uint16_t led);
+void Leds_Off(uint16_t led);
 void Leds_AllOn(void);
 void Leds_AllOff(void);
